Starting with the refactoring and cleaning the code
