+++ src/styles/Modal.module.css
@@ -14,16 +14,14 @@
   background-color: white;
   width: 100%;
   margin: auto;
-  margin-top: 6rem;
   border-radius: 5px;
-  height: 90%;
   overflow: scroll;
   position: relative;
 }
 
 @media(min-width: 577px) {
   .details {
-    margin-top: 4rem;
+    margin-top: 2rem;
     height: auto;
   }
 }
